bgpd: widen per-AF flags to 32 bits so 'install backup-path' can be disabled
BGP_CONFIG_BACKUP_PATH_FLUSH is (1 << 16), but struct bgp's
af_flags[][] is a uint16_t.  SET_FLAG()/CHECK_FLAG() for bit 16 are
therefore silently truncated: the flush flag is never actually stored.

As a result, on 'no install backup-path' the flush is a no-op --
bgp_clear_backup_paths() never runs, BGP_PATH_BACKUP_CHG is never set,
and the already-installed backup nexthops are never withdrawn from
zebra, so the FIB keeps a stale backup.

Widen af_flags to uint32_t so the flush bit (and any future per-AF flag
past bit 15) is stored correctly.

+++ bgpd/bgpd.h
@@ -818,7 +818,7 @@ struct bgp {
 	bool gr_route_sync_pending;
 
 	/* BGP Per AF flags */
-	uint16_t af_flags[AFI_MAX][SAFI_MAX];
+	uint32_t af_flags[AFI_MAX][SAFI_MAX];
 #define BGP_CONFIG_DAMPENING				(1 << 0)
 /* l2vpn evpn flags - 1 << 0 is used for DAMPENNG */
 #define BGP_L2VPN_EVPN_ADV_IPV4_UNICAST (1 << 1)
